+++ src/math/operations/cutPasteOps/gyrate.ts
@@ -1,8 +1,7 @@
 import { withOrigin } from "math/geom"
 import { Polyhedron } from "math/polyhedra"
 import { mapObject } from "utils"
-import Capstone from "data/specs/Capstone"
-import Composite from "data/specs/Composite"
+import { Capstone, Composite } from "data/specs"
 import { CapOptions, capOptionArgs } from "./cutPasteUtils"
 import { getTransformedVertices } from "../operationUtils"
 import { makeOperation } from "../Operation"
