+++ src/math/operations/operationUtils.ts
@@ -9,9 +9,7 @@ import {
   withOrigin,
 } from "math/geom"
 import { mapObject } from "utils"
-import PolyhedronSpecs from "data/specs/PolyhedronSpecs"
-import { Facet } from "data/specs/Classical"
-import { Twist } from "types"
+import { PolyhedronSpecs, Facet, Twist } from "data/specs"
 
 export interface FacetOpts {
   facet: Facet
